modificaciones al proyecto

+++ src/main/java/sv/edu/ues/occ/ingenieria/prn335/inventario/web/boundary/rest/CompraResource.java
@@ -158,7 +158,7 @@ public class CompraResource {
                         entity.setProveedor(proveedor);
                     }
 
-                    entity.setId(Math.toIntExact(id));
+                    entity.setId((long) Math.toIntExact(id));
                     compraDao.modificar(entity);
                     return Response.ok(entity).build();
                 }
